feat: renamed llm_service to ll_chatbot
